refactor(flterm): split glyph rasterizer lanes
