+++ packages/flterm/test/rendering/atlas/glyph_text_atlas_lane_test.dart
@@ -13,7 +13,7 @@ void main() {
 
     setUp(() {
       rasterizer = GlyphRasterizer()..configure(_config());
-      lane = GlyphTextAtlasLane(rasterizer);
+      lane = GlyphTextAtlasLane(rasterizer.textRasterizer);
     });
 
     tearDown(() => rasterizer.dispose());
